+++ app/Providers/AppServiceProvider.php
@@ -4,18 +4,17 @@ declare(strict_types=1);
 
 namespace App\Providers;
 
+use Illuminate\Pagination\Paginator;
 use Illuminate\Support\ServiceProvider;
 
 class AppServiceProvider extends ServiceProvider
 {
-
     public function register(): void
     {
-
     }
 
     public function boot(): void
     {
-
+        Paginator::useBootstrapFive();
     }
 }